refactor path sale controller
